+++ pajbot/bot.py
@@ -51,11 +51,6 @@ from pajbot import utils
 
 log = logging.getLogger(__name__)
 
-URL_REGEX = re.compile(
-    r"\(?(?:(http|https):\/\/)?(?:((?:[^\W\s]|\.|-|[:]{1})+)@{1})?((?:www.)?(?:[^\W\s]|\.|-)+[\.][^\W\s]{2,4}|localhost(?=\/)|\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})(?::(\d*))?([\/]?[^\s\?]*[\/]{1})*(?:\/?([^\s\n\?\[\]\{\}\#]*(?:(?=\.)){1}|[^\s\n\?\[\]\{\}\.\#]*)?([\.]{1}[^\s\?\#]*)?)?(?:\?{1}([^\s\n\#\[\]]*))?([\#][^\s\n]*)?\)?",
-    re.IGNORECASE,
-)
-
 SLICE_REGEX = re.compile(r"(-?\d+)?(:?(-?\d+)?)?")
 
 
@@ -389,6 +384,24 @@ class Bot:
 
         return None
 
+    def get_date_value(self, key, extra={}):
+        try:
+            tz = timezone(key)
+            return datetime.datetime.now(tz).strftime("%Y-%m-%d")
+        except:
+            log.exception("Unhandled exception in get_date_value")
+
+    def get_datetimefromisoformat_value(self, key, extra={}):
+        try:
+            dt = datetime.datetime.fromisoformat(key)
+            if dt.tzinfo is None:
+                # The date format passed through in key did not contain a timezone, so we replace it with UTC
+                dt = dt.replace(tzinfo=datetime.timezone.utc)
+
+            return dt
+        except:
+            log.exception("Unhandled exception in get_datetimefromisoformat_value")
+
     def get_current_song_value(self, key, extra={}):
         if self.stream_manager.online:
             current_song = PleblistManager.get_current_song(self.stream_manager.current_stream.id)
@@ -526,6 +539,15 @@ class Bot:
         diff = now - molly_birth
         return diff.total_seconds() / 3600 / 24 / 365
 
+    def get_datetime_value(self, key, extra=[]):
+        try:
+            tz = timezone(key)
+            return datetime.datetime.now(tz)
+        except:
+            log.exception("Unhandled exception in get_datetime_value")
+
+        return None
+
     @property
     def is_online(self):
         return self.stream_manager.online
@@ -625,15 +647,6 @@ class Bot:
         else:
             log.warning("Unknown send_message method: %s", method)
 
-    def safe_privmsg(self, message, channel=None):
-        # Check for banphrases
-        res = self.banphrase_manager.check_message(message, None)
-        if res is not False:
-            self.privmsg(f"filtered message ({res.id})", channel)
-            return
-
-        self.privmsg(message, channel)
-
     def say(self, message, channel=None):
         if message is None:
             log.warning("message=None passed to Bot::say()")
@@ -646,11 +659,20 @@ class Bot:
         self.privmsg(message[:510], channel)
 
     def is_bad_message(self, message):
+        # Checks for banphrases
         return self.banphrase_manager.check_message(message, None) is not False
+
+    def safe_privmsg(self, message, channel=None):
+        if not self.is_bad_message(message):
+            self.privmsg(message, channel)
 
     def safe_me(self, message, channel=None):
         if not self.is_bad_message(message):
             self.me(message, channel)
+
+    def safe_say(self, message, channel=None):
+        if not self.is_bad_message(message):
+            self.say(message, channel)
 
     def me(self, message, channel=None):
         self.say("/me " + message[:500], channel=channel)
@@ -802,7 +824,7 @@ class Bot:
 
         with DBManager.create_session_scope(expire_on_commit=False) as db_session:
             source = User.from_basics(db_session, UserBasics(id, login, name))
-            res = HandlerManager.trigger("on_pubmsg", source=source, message=event.arguments[0])
+            res = HandlerManager.trigger("on_pubmsg", source=source, message=event.arguments[0], tags=tags)
             if res is False:
                 return False
 
@@ -915,6 +937,7 @@ class Bot:
             else utils.time_since(var * 60, 0, time_format="long"),
             "time_since": lambda var, args: "no time" if var == 0 else utils.time_since(var, 0, time_format="long"),
             "time_since_dt": _filter_time_since_dt,
+            "timedelta_days": _filter_timedelta_days,
             "urlencode": _filter_urlencode,
             "join": _filter_join,
             "number_format": _filter_number_format,
@@ -923,6 +946,11 @@ class Bot:
             "or_broadcaster": self._filter_or_broadcaster,
             "or_streamer": self._filter_or_broadcaster,
             "slice": _filter_slice,
+            "subtract": _filter_subtract,
+            "multiply": _filter_multiply,
+            "divide": _filter_divide,
+            "floor": _filter_floor,
+            "ceil": _filter_ceil,
         }
         if f.name in available_filters:
             return available_filters[f.name](resp, f.arguments)
@@ -934,7 +962,7 @@ class Bot:
     def find_unique_urls(self, message):
         from pajbot.modules.linkchecker import find_unique_urls
 
-        return find_unique_urls(URL_REGEX, message)
+        return find_unique_urls(message)
 
 
 def _filter_time_since_dt(var, args):
@@ -946,6 +974,14 @@ def _filter_time_since_dt(var, args):
         return "0 seconds"
     except:
         return "never FeelsBadMan ?"
+
+
+def _filter_timedelta_days(var, args):
+    try:
+        td = utils.now() - var
+        return str(td.days)
+    except:
+        return "0"
 
 
 def _filter_join(var, args):
@@ -979,7 +1015,58 @@ def lowercase_first_letter(s):
 
 def _filter_add(var, args):
     try:
-        return str(int(var) + int(args[0]))
+        lh = utils.parse_number_from_string(var)
+        rh = utils.parse_number_from_string(args[0])
+
+        return str(lh + rh)
+    except:
+        return ""
+
+
+def _filter_subtract(var, args):
+    try:
+        lh = utils.parse_number_from_string(var)
+        rh = utils.parse_number_from_string(args[0])
+
+        return str(lh - rh)
+    except:
+        return ""
+
+
+def _filter_multiply(var, args):
+    try:
+        lh = utils.parse_number_from_string(var)
+        rh = utils.parse_number_from_string(args[0])
+
+        return str(lh * rh)
+    except:
+        return ""
+
+
+def _filter_divide(var, args):
+    try:
+        lh = utils.parse_number_from_string(var)
+        rh = utils.parse_number_from_string(args[0])
+
+        return str(lh / rh)
+    except:
+        return ""
+
+
+def _filter_floor(var, args):
+    import math
+
+    try:
+        return str(math.floor(float(var)))
+    except:
+        return ""
+
+
+def _filter_ceil(var, args):
+    import math
+
+    try:
+        return str(math.ceil(float(var)))
     except:
         return ""