+++ pajbot/modules/linkchecker.py
@@ -5,6 +5,7 @@ import urllib.parse
 import requests
 from bs4 import BeautifulSoup
 from sqlalchemy import Column, INT, TEXT
+from urlextract import URLExtract
 
 import pajbot.managers
 import pajbot.models
@@ -20,6 +21,10 @@ from pajbot.modules import BaseModule
 from pajbot.modules import ModuleSetting
 
 log = logging.getLogger(__name__)
+
+
+extractor = URLExtract()
+extractor.update_when_older(14)
 
 
 def is_subdomain(x, y):
@@ -61,11 +66,9 @@ def is_same_url(x, y):
     )
 
 
-def find_unique_urls(regex, message):
-    _urls = regex.finditer(message)
+def find_unique_urls(message):
     urls = []
-    for i in _urls:
-        url = i.group(0)
+    for url in extractor.gen_urls(message):
         if not (url.startswith("http://") or url.startswith("https://")):
             url = "http://" + url
         if not (url[-1].isalpha() or url[-1].isnumeric() or url[-1] == "/"):
@@ -162,6 +165,7 @@ class LinkCheckerModule(BaseModule):
         ModuleSetting(
             key="ban_sub_links", label="Disallow links from subscribers", type="boolean", required=True, default=False
         ),
+        ModuleSetting(key="vip_exemption", label="Allow links from VIPs", type="boolean", required=True, default=False),
         ModuleSetting(
             key="timeout_length",
             label="Timeout length",
@@ -273,6 +277,9 @@ class LinkCheckerModule(BaseModule):
             return
 
         if source.level >= self.settings["bypass_level"] or source.moderator is True:
+            return
+
+        if self.settings["vip_exemption"] and source.vip is True:
             return
 
         if len(urls) > 0: