+++ pajbot/modules/repspam.py
@@ -17,6 +17,14 @@ class RepspamModule(BaseModule):
     CATEGORY = "Moderation"
     SETTINGS = [
         ModuleSetting(
+            key="enabled_by_stream_status",
+            label="Enable moderation of repetitive spam when the stream is:",
+            type="options",
+            required=True,
+            default="Offline and Online",
+            options=["Online Only", "Offline Only", "Offline and Online"],
+        ),
+        ModuleSetting(
             key="bypass_level",
             label="Level to bypass module",
             type="number",
@@ -99,8 +107,15 @@ class RepspamModule(BaseModule):
         return len(word) <= 0
 
     def on_message(self, source, message, whisper, **rest):
+        if self.settings["enabled_by_stream_status"] == "Online Only" and not self.bot.is_online:
+            return
+
+        if self.settings["enabled_by_stream_status"] == "Offline Only" and self.bot.is_online:
+            return
+
         if whisper:
             return
+
         if source.level >= self.settings["bypass_level"] or source.moderator:
             return