+++ CHANGELOG.md
@@ -2,32 +2,65 @@
 
 ## Unversioned
 
+## v1.50
+
+Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` when updating to this version!
+
+- Major: Math filters (`add`, `subtract`, `multiply`, and `divide`) are now able to read float values in addition to int values.
+- Minor: Added support for the `delete` moderation action in the ASCII module. (#1174)
+- Minor: Added an option to the `stream_update` module that allows moderators to change the title/game (without the level requirement). (#1165)
+- Minor: Added a new module to print a chat/whsiper alert on cheer. (#1158)
+- Minor: Added the filter `timedelta_days` which returns the amount of days between now and a `datetime` object. (#1173)
+- Minor: Added the filter `ceil` which returns the smallest integer greater than or equal to the parameter. (#1168)
+- Minor: Added the filter `floor` which returns the largest integer less than or equal to the parameter. (#1168)
+- Minor: Added the variable `datetimefromisoformat` which allows commands to generate a full datetime object from a given string, which can then be further expanded on using filters. (#1169)
+- Minor: Filter arguments now allow the period character `.` (#1171)
+- Bugfix: Escape message content in command examples. (#1181)
+
+## v1.49
+
+Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` when updating to this version!
+
+- Major: `$(urlfetch)` now handles non-200 error codes differently by returning the body if the returned content type is plain text, or a generic "urlfetch error 404" if the content type was not plain text. (#1140)
+- Minor: Added a VIP exemption option to the case checker module. (#1150)
+- Minor: Added a VIP exemption option to the link checker module. (#1149)
+- Minor: Added an option to only enable slots after a re/sub. (#1146)
+- Minor: Added an option to allow ASCII characters while the streamer is on/offline. (#1145)
+- Minor: Added an option to allow repetitive spam while the streamer is on/offline. (#1144)
+- Minor: Added an option to allow mass pings while the streamer is on/offline. (#1129)
+- Minor: Added an option to allow rouletting while the streamer is on/offline. (#1131)
+- Minor: Added `subtract`, `multiply` & `divide` filters. (#1136)
+- Minor: Added `$(datetime:<timezone>)` variable. This allows users to parse their own timezone's time and use the `strftime` filter. (#1132)
+- Minor: Added `$(date:<timezone>)` variable. (#1125)
+- Bugfix: Fix issue where we matched some normal messages as links, e.g. 1.40 or asd...xd (#1148)
+
+## v1.48
+
 Because the Game/Title setting API calls are now using the Helix calls, it's no longer possible to use the Bot token to update the game/title of a channel, instead the Streamer token **must** be used. In addition to this, the Streamer token needs a new permission `user:edit:broadcast`.  
 In short: The Streamers must re-authenticate with the `/streamer_login` endpoint for `!setgame` and `!settitle` to work.
 
 Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` when updating to this version!
 
 - Breaking: Replace Kraken Game/Title setting API calls with Helix ones. (#1001)
+- Minor: Added user-specific cooldown to playsound module (#888, #1006)
+- Minor: The permaban module will now ban immediately on command use (#1014)
 - Minor: Bot now provides more helpful error message when `!clip` is used in a channel with clips disabled. (#1091)
 - Minor: Added custom message options to the LastFM module. (#1090)
 - Minor: Added customizable cooldowns to the LastFM module. (#1090)
 - Minor: Added an online only option to the LastFM module. (#1090)
 - Minor: Added option to select your referred gambling command. (#1067)
 - Minor: The `sub_only` command option will now show in the `!debug command` command. (#1027)
-- Minor: The permaban module will now ban immediately on command use (#1014)
 - Minor: Added settings to change command name and cooldowns for showemote module. (#1007)
 - Minor: Replace Kraken Game/Title/Stream/Video fetching API calls with Helix ones (#1001)
 - Minor: Removed excess message in whisper for paid timeout module. (#993)
 - Minor: Move the streamer image resizing to the css file. Also added a rounded border to it. (#992)
-- Minor: Added user-specific cooldown to playsound module (#888, #1006)
 - Minor: Add \$(broadcaster) variable (#925, #1076)
 - Minor: Duel winrate now takes `duels_won` into consideration if winrate is equal (#1079)
 - Minor: Moved CLR-based modules to a sub-folder. (#1094)
-- Bugfix: Paid timeouts will now only timeout once. (#993)
-- Bugfix: Handle new format of P&SL lists (#988, #989)
 - Bugfix: Fixed league rank module not working at all. (#990)
-- Bugfix: Handle new format of P&SL lists (#988, #989, #994)
+- Bugfix: Paid timeouts will now only timeout once. (#993)
 - Bugfix: Fixed name of "get timed out" quest. (It just said "Quest" before) (#1003)
+- Bugfix: Handle new format of P&SL lists (#988, #989, #994)
 - Bugfix: Added proper error handling to P&SL module. (#991, #1005)
 
 ## v1.47